Add tutorial for developing interface features
This tutorial covers how to plumb data, service calls, etc. from
the ROS world through WebRTC and up to the web interface, where
a React component (i.e. a piece of UI that the user can interact
with) can display data, trigger services, etc. It takes the
example of homing the robot, which involves showing a button
that calls the /home_the_robot ROS2 service, showing a loading
spinner while the robot is doing the homing sequence (as reported
by the driver on the /mode ROS2 topic), and disappearing once
the robot is done (as reported on the /is_homed topic).