Remove numpy pin in Web Teleop (#97)
* Remove numpy pin in Web Teleop

This commits removes the pin on an older version of numpy. Stretch Body now pins the right version of numpy. Details in these two PRs:
 - https://github.com/hello-robot/stretch_body/pull/349
 - https://github.com/hello-robot/stretch_body/pull/348

* Numpy primitives deprecated

With numpy 1.24 and onwards, using np.bool and other np.<primitive>
types is deprecated, and raises exceptions. We are supposed to use
the built-in primitives (i.e. bool, float, etc.).